fixed warnings
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2857 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/laser2d/mainloop.cpp
@@ -152,8 +152,8 @@ MainLoop::readData( orca::LaserScanner2dDataPtr &laserData )
     driver_->read( laserData->ranges, 
                    laserData->intensities,
                    laserData->timeStamp );
-    assert( laserData->ranges.size()      == config_.numberOfSamples );
-    assert( laserData->intensities.size() == config_.numberOfSamples );
+    assert( laserData->ranges.size()      == (unsigned int)config_.numberOfSamples );
+    assert( laserData->intensities.size() == (unsigned int)config_.numberOfSamples );
 
     // flip the scan left-to-right if we are configured to do so
     if ( compensateRoll_ ) {
